Created and finished my keylogger
This is all of the code necessary for your key logger just change the directory where your text files are and it should work like a charm. This isn't super fast so it will mess up especially the longer you run the program but it works decently. I might make the code more efficient in the future.